# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Ambulanzsystem V1 API
  version: 0.1.0
  description: |
    Frozen Stage-0 contract for the Ambulanzsystem V1 (see docs/plan/master-plan.md).
    Locked decisions D1-D8 apply. JSON property names are serialized exactly as
    written here (no automatic camelCase transformation server-side).

    Auth token types (JWT `type` claim): `admin`, `leitstelle`, `user`, `qr`.
    Policies referenced below:
    - AdminOnly: type=admin
    - LeitstelleOrAdmin: type in (admin, leitstelle)
    - AuthenticatedUser: type in (admin, leitstelle, user)
    - TriageWrite: type in (admin, leitstelle, user, qr)

    Canonical triage values: rot | gelb | gruen | schwarz (ASCII, D5).
    `blau` and Unicode `grün` are rejected at the write boundary.
servers:
  - url: http://localhost:5000
    description: local backend
  - url: http://localhost:4010
    description: Prism mock server
security:
  - bearerAuth: []
...
paths:
...
  /api/persons:
    get:
      tags: [persons]
      summary: List patients linked to a scene (situation room + patient overview). Policy TriageWrite. Every request is read-audit-logged (D7).
      parameters:
        - name: operationSceneId
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Patients.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Patient" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/verify-patient-qr-code:
    post:
      tags: [persons]
      summary: >
        Scan a patient QR tag. Unknown token -> 404. Unbound -> creates Patient +
        fully-seeded Body, binds code, links scene (201). Bound -> returns existing
        patient, refreshes scene link (200). Concurrency-safe (row lock) — parallel
        scans of the same code yield the same patient. Online-only. Policy TriageWrite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qr_code, operationSceneId]
              properties:
                qr_code: { type: string, maxLength: 128 }
                operationSceneId: { type: integer }
      responses:
        "200":
          description: Existing patient returned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyQrResult" }
        "201":
          description: New patient created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyQrResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: Unknown QR token — rejected (UC-04). }
  /api/persons/manual:
    post:
      tags: [persons]
      summary: >
        Create a patient without a QR code. Idempotent on clientGeneratedId (D3):
        replaying the same UUID returns the already-created patient (200) instead of
        a duplicate (201 on first create). Server generates humanReadableId
        (no patient-identifying content). Policy TriageWrite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operationSceneId]
              properties:
                operationSceneId: { type: integer }
                name: { type: string }
                clientGeneratedId: { type: string, format: uuid, description: Offline idempotency key generated by the client. }
      responses:
        "200":
          description: Idempotent replay — patient already existed for this clientGeneratedId.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/persons/{id}/update-triage-color:
    post:
      tags: [persons]
      summary: >
        Record-only START triage write (D1). triageColor validated against
        rot|gelb|gruen|schwarz (shared normalize helper maps 'grün'->'gruen', trims,
        lowercases; everything else incl. 'blau' -> 400). Field-level LWW merge using
        clientUpdatedAt when supplied (NFR-SAFE-08/09). Audit-logged with
        before/after. Publishes realtime update. Policy TriageWrite.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                triageColor: { $ref: "#/components/schemas/TriageColor" }
                respiration: { type: boolean }
                blutung: { type: boolean }
                radialispuls: { type: boolean }
                transport: { type: boolean }
                dringend: { type: boolean }
                kontaminiert: { type: boolean }
                clientUpdatedAt: { type: string, format: date-time, description: Client-side write timestamp for offline replay merge. }
      responses:
        "200":
          description: Updated patient.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "400": { description: Invalid triage value (rejected at write boundary, FR-TRIAGE-04). }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/persons/{id}/respiration:
    post:
      tags: [persons]
      summary: Set the breathing-sufficient flag. Publishes realtime update. Policy TriageWrite.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [respiration]
              properties:
                respiration: { type: boolean }
                clientUpdatedAt: { type: string, format: date-time }
      responses:
        "200":
          description: Updated patient.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/persons/{id}/location:
    post:
      tags: [persons]
      summary: Update patient position (auto-captured device GPS or manual correction, FR-LOC-05/06). Publishes realtime update. Policy TriageWrite.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat: { type: number, minimum: -90, maximum: 90 }
                lng: { type: number, minimum: -180, maximum: 180 }
                source: { type: string, enum: [gps, manual], default: gps }
                accuracyMeters: { type: number, description: Reported GPS accuracy; UI flags >10m as unreliable (FR-LOC-07). }
                indoorLocation: { type: string, description: "Optional free text: tent/sector/room/floor/treatment bay (FR-LOC-08)." }
                clientUpdatedAt: { type: string, format: date-time, description: Client-side write timestamp for offline replay merge (NFR-SAFE-08/09), consistent with update-triage-color. }
      responses:
        "200":
          description: Updated patient.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/persons/{id}/reassign-qr-code:
    post:
      tags: [persons]
      summary: Unbind the patient's current QR code and bind the given UNBOUND code (mis-attached tag correction, FR-QR-07). Policy TriageWrite.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qr_code]
              properties:
                qr_code: { type: string, maxLength: 128 }
      responses:
        "200":
          description: Updated patient with new QR binding.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Patient" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Given code is already bound to another patient. }
  /api/persons/{id}/triage-history:
    get:
      tags: [persons]
      summary: Triage change history for a patient, sourced from the audit log (FR-SIT-08). Policy LeitstelleOrAdmin.
      parameters: [{ $ref: "#/components/parameters/idPath" }]
      responses:
        "200":
          description: Chronological triage changes.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TriageHistoryEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ------------------------------------------------- ambulanzprotokoll
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    Unauthorized:
      description: Missing/invalid/expired/revoked credentials. Expired access blocks new actions and reads (FR-AUTH-09).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Entity not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [status, message]
      properties:
        status: { type: string, const: error }
        message: { type: string }
...
    TriageColor:
      type: string
      enum: [rot, gelb, gruen, schwarz]
      description: Canonical ASCII values (D5). UI displays 'grün' for 'gruen'. 'blau' is invalid in V1.
...
    Patient:
      type: object
      required: [id, operationSceneId, createdAt, updatedAt]
      properties:
        id: { type: integer }
        humanReadableId: { type: [string, "null"], description: Set for manually created patients; no patient-identifying content (FR-PAT-07). }
        clientGeneratedId: { type: [string, "null"], format: uuid }
        name: { type: [string, "null"] }
        triagefarbe:
          oneOf:
            - $ref: "#/components/schemas/TriageColor"
            - type: "null"
        atmung: { type: [boolean, "null"] }
        blutung: { type: [boolean, "null"] }
        radialispuls: { type: [boolean, "null"] }
        transport: { type: [boolean, "null"] }
        dringend: { type: [boolean, "null"] }
        kontaminiert: { type: [boolean, "null"] }
        longitudePatient: { type: [number, "null"] }
        latitudePatient: { type: [number, "null"] }
        locationSource: { type: [string, "null"], enum: [gps, manual, null] }
        locationAccuracyMeters: { type: [number, "null"] }
        indoorLocation: { type: [string, "null"] }
        locationUpdatedAt: { type: [string, "null"], format: date-time }
        operationSceneId: { type: integer }
        qrCodeBound: { type: boolean, description: Whether a patient QR tag is currently bound. }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    VerifyQrResult:
      type: object
      required: [patient, created]
      properties:
        patient: { $ref: "#/components/schemas/Patient" }
        created: { type: boolean, description: True when this scan created the patient. }
    TriageHistoryEntry:
      type: object
      required: [timestamp, actorId, field, before, after]
      properties:
        timestamp: { type: string, format: date-time }
        actorId: { type: integer }
        actorRole: { type: string, enum: [admin, leitstelle, user, qr] }
        field: { type: string, description: e.g. triagefarbe, atmung, blutung. }
        before: { type: [string, "null"] }
        after: { type: [string, "null"] }
...